Move tests to the right test module
