+++ tests/unit/states/test_saltmod.py
@@ -4,6 +4,7 @@
 '''
 # Import Python libs
 from __future__ import absolute_import
+import os
 import time
 import tempfile
 
@@ -204,3 +205,57 @@ class SaltmodTestCase(TestCase, LoaderModuleMockMixin):
 
         with patch.dict(saltmod.__salt__, {'saltutil.wheel': wheel_mock}):
             self.assertDictEqual(saltmod.wheel(name), ret)
+
+
+@skipIf(NO_MOCK, NO_MOCK_REASON)
+class StatemodTests(TestCase, LoaderModuleMockMixin):
+    def setup_loader_modules(self):
+        self.tmp_cachedir = tempfile.mkdtemp(dir=TMP)
+        return {
+            saltmod: {
+                '__env__': 'base',
+                '__opts__': {
+                    'id': 'webserver2',
+                    'argv': [],
+                    '__role': 'master',
+                    'cachedir': self.tmp_cachedir,
+                    'extension_modules': os.path.join(self.tmp_cachedir, 'extmods'),
+                },
+                '__salt__': {'saltutil.cmd': MagicMock()},
+                '__orchestration_jid__': salt.utils.jid.gen_jid()
+            }
+        }
+
+    def test_statemod_state(self):
+        ''' Smoke test for for salt.states.statemod.state().  Ensures that we
+            don't take an exception if optional parameters are not specified in
+            __opts__ or __env__.
+        '''
+        args = ('webserver_setup', 'webserver2')
+        kwargs = {
+            'tgt_type': 'glob',
+            'fail_minions': None,
+            'pillar': None,
+            'top': None,
+            'batch': None,
+            'orchestration_jid': None,
+            'sls': 'vroom',
+            'queue': False,
+            'concurrent': False,
+            'highstate': None,
+            'expr_form': None,
+            'ret': '',
+            'ssh': False,
+            'timeout': None, 'test': False,
+            'allow_fail': 0,
+            'saltenv': None,
+            'expect_minions': False
+        }
+        ret = saltmod.state(*args, **kwargs)
+        expected = {
+            'comment': 'States ran successfully.',
+            'changes': {},
+            'name': 'webserver_setup',
+            'result': True
+        }
+        self.assertEqual(ret, expected)
